tiktok-scraper/requirements-2.txt
appdirs==1.4.4
certifi==2020.6.20
chardet==3.0.4
idna==2.10
numpy==1.19.2
pandas==1.1.3
pyee==7.0.4
pyppeteer==0.2.2
pyppeteer-stealth==2.6.0
python-dateutil==2.8.1
PyTikTokAPI==0.0.4
pytz==2020.1
requests==2.24.0
six==1.15.0
tqdm==4.50.1
urllib3==1.25.10
websockets==8.1
